+++ src/_pytask/database.py
@@ -74,26 +74,29 @@ def pytask_extend_command_line_interface(cli: click.Group) -> None:
             type=click.Choice([x.value for x in _DatabaseProviders]),
             help=(
                 "Database provider. All providers except sqlite are considered "
-                "experimental.  [default: sqlite]"
+                "experimental.  [dim]\\[default: sqlite][/]"
             ),
             default=None,
         ),
         click.Option(
             ["--database-filename"],
             type=click.Path(),
-            help="Path to database relative to root.  [default: .pytask.sqlite3]",
+            help=(
+                "Path to database relative to root.  "
+                "[dim]\\[default: .pytask.sqlite3][/]"
+            ),
             default=None,
         ),
         click.Option(
             ["--database-create-db"],
             type=bool,
-            help="Create database if it does not exist.  [default: True]",
+            help="Create database if it does not exist.  [dim]\\[default: True][/]",
             default=None,
         ),
         click.Option(
             ["--database-create-tables"],
             type=bool,
-            help="Create tables if they do not exist.  [default: True]",
+            help="Create tables if they do not exist.  [dim]\\[default: True][/]",
             default=None,
         ),
     ]